Renamed files and split autonomous into two files - one for red team, one for blue team.
